requirements.txt
# Core dependencies
Pillow>=10.0.0
python-dotenv>=1.0.1
openai>=1.13.3
pandas>=2.0.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
